requirements.txt
astroid==2.4.2
click==7.1.2
fastapi==0.61.1
flake8==3.8.4
h11==0.11.0
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
motor==2.3.0
pycodestyle==2.6.0
pydantic==1.6.1
pydocstyle==5.1.1
pyflakes==2.2.0
pylint==2.6.0
pymongo==3.11.0
python-decouple==3.3
six==1.15.0
snowballstemmer==2.0.0
starlette==0.13.6
style==1.1.0
toml==0.10.1
update==0.0.1
uvicorn==0.12.2
wrapt==1.12.1
